requirements.txt
pywebview>=4.2.2
pyinstaller>=5.13.0
Pillow>=10.0.0